backend/pyproject.toml
[project]
name = "bafain-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
  "fastapi",
  "uvicorn[standard]",
  "supabase",
  "redis",
  "python-dotenv",
]
